docs: sendpay_failure timestamps are u32
sendpay_failure serializes the payment via
json_add_payment_fields(), which emits created_at and
completed_at with json_add_u32 (struct wallet_payment keeps them
as u32):

  lightningd/pay.c:158,160
  wallet/wallet.h:363-364

Changelog-None

+++ cln-rpc/src/notifications.rs
@@ -725,9 +725,9 @@ pub struct SendpayFailureData {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub bolt12: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub completed_at: Option<u64>,
+    pub completed_at: Option<u32>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<u64>,
+    pub created_at: Option<u32>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub created_index: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
